# api/openapi.yaml
openapi: 3.1.0
info:
  title: sing-box-panel API
  version: 0.1.0
  license:
    name: GPL-3.0-or-later
    identifier: GPL-3.0-or-later
servers:
  - url: /api/v1
paths:
...
  /config/render:
    post:
      operationId: renderStructuredConfiguration
      description: Projects the current canonical head through one exact pinned declarative capability, persists immutable startup bytes, and queues their exact-binary check. When core_version is omitted, the verified running exact version and artifact are used; latest is never substituted.
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StructuredRenderRequest" }
      responses:
        "202":
          description: The pending immutable structured candidate and its durable startup-check task.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StructuredRender" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /config/artifacts:
    get:
      operationId: listStartupArtifacts
      description: Lists durable structured and manual startup candidates without exposing their potentially secret configuration bytes.
      parameters:
        - name: canonical_revision_id
          in: query
          schema: { type: string, minLength: 1, maxLength: 256 }
        - name: core_version
          in: query
          schema: { $ref: "#/components/schemas/ExactCoreVersion" }
        - name: core_artifact_id
          in: query
          schema: { type: string, minLength: 1, maxLength: 256 }
        - name: kind
          in: query
          schema: { type: string, enum: [structured, manual] }
        - name: state
          in: query
          schema: { $ref: "#/components/schemas/ManualArtifactState" }
        - name: before_time
          in: query
          description: RFC 3339 cursor time; must be paired with before_id.
          schema: { type: string, format: date-time }
        - name: before_id
          in: query
          description: Stable cursor ID; must be paired with before_time.
          schema: { type: string, minLength: 1, maxLength: 256 }
        - $ref: "#/components/parameters/PageLimit"
      responses:
        "200":
          description: A newest-first page of immutable startup-candidate metadata.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StartupArtifactPage" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
...
  /config/manual/{startupArtifactId}/reattach/preview:
    parameters:
      - $ref: "#/components/parameters/StartupArtifactID"
    get:
      operationId: previewManualReattach
      description: Recomputes a read-only base/current/manual three-way preview using the exact currently pinned structured manifest.
      responses:
        "200":
          description: Immutable evidence, reversible owned partial, residual paths, and explicit conflicts.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ManualReattachPreview" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
  /config/manual/{startupArtifactId}/reattach:
    parameters:
      - $ref: "#/components/parameters/StartupArtifactID"
    post:
      operationId: applyManualReattach
      description: Recomputes the preview, verifies all evidence, requires one explicit choice per conflict, and atomically creates a new revision, manual artifact, and startup-check task.
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ManualReattachApplyRequest" }
      responses:
        "202":
          description: Newly committed canonical revision and immutable manual candidate.
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ManualReattachSave" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    managementBearer:
      type: http
      scheme: bearer
    sessionCookie:
      type: apiKey
      in: cookie
      name: sbp_session
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required for cookie-authenticated state-changing requests.
      schema:
        type: string
        minLength: 1
...
    StartupArtifactID:
      name: startupArtifactId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 256 }
...
    PageLimit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
    Problem:
      description: Stable RFC 9457-compatible problem response.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        request_id: { type: string }
        fields:
          type: object
          additionalProperties: { type: string }
    CanonicalDocument:
      type: object
      additionalProperties: false
      required: [schema_version, global, nodes, rules, subscription]
      properties:
        schema_version: { type: integer, const: 1 }
        global: { type: object, additionalProperties: true }
        nodes:
          type: array
          maxItems: 10000
          items:
            type: object
            required: [id, kind, enabled]
            properties:
              id: { type: string }
              kind: { type: string }
              enabled: { type: boolean }
            additionalProperties: true
        rules:
          type: array
          maxItems: 10000
          items:
            type: object
            required: [id, enabled]
            properties:
              id: { type: string }
              enabled: { type: boolean }
            additionalProperties: true
        subscription: { type: object, additionalProperties: true }
    CanonicalSnapshot:
      type: object
      additionalProperties: false
      required: [id, sequence, schema_version, document, document_json, sha256, created_at]
      properties:
        id: { type: string }
        sequence: { type: integer, format: int64, minimum: 1 }
        parent_id: { type: string }
        schema_version: { type: integer, const: 1 }
        document: { $ref: "#/components/schemas/CanonicalDocument" }
        document_json:
          type: string
          description: Exact canonical JSON bytes encoded as a JSON string for lossless browser parsing.
        sha256: { type: string, pattern: "^[0-9a-f]{64}$" }
        created_at: { type: string, format: date-time }
...
    ExactCoreVersion:
      type: string
      description: A non-zero exact stable semantic version without a leading v, prerelease, or build metadata.
      maxLength: 62
      pattern: "^(0|[1-9][0-9]*)\\.(0|[1-9][0-9]*)\\.(0|[1-9][0-9]*)$"
      not: { const: "0.0.0" }
    SHA256:
      type: string
      pattern: "^[0-9a-f]{64}$"
    GitObjectID:
      type: string
      pattern: "^([0-9a-f]{40}|[0-9a-f]{64})$"
...
    CapabilitySupportLevel:
      type: string
      enum: [native_structured, compatible_structured, manual_json, unavailable]
...
    RuntimeCoreIdentity:
      type: object
      additionalProperties: false
      required:
        - pid
        - process_start_token
        - exact_core_version
        - core_artifact_id
        - archive_sha256
        - binary_sha256
        - activation_bundle_id
      properties:
        pid: { type: integer, minimum: 1 }
        process_start_token: { type: string }
        exact_core_version: { $ref: "#/components/schemas/ExactCoreVersion" }
        core_artifact_id: { type: string }
        archive_sha256: { $ref: "#/components/schemas/SHA256" }
        binary_sha256: { $ref: "#/components/schemas/SHA256" }
        activation_bundle_id: { type: string }
    CoreVersionResolution:
      type: object
      additionalProperties: false
      required: [exact_version, source]
      properties:
        exact_version: { $ref: "#/components/schemas/ExactCoreVersion" }
        source:
          type: string
          enum: [explicit, running]
        running: { $ref: "#/components/schemas/RuntimeCoreIdentity" }
    CapabilityPin:
      type: object
      additionalProperties: false
      description: Persisted immutable capability reference.
      required: [exact_core_version, repository, commit_sha, manifest_sha256, support_level, pinned_at]
      properties:
        exact_core_version: { $ref: "#/components/schemas/ExactCoreVersion" }
        repository: { type: string, pattern: "^[^/]+/[^/]+$" }
        commit_sha: { $ref: "#/components/schemas/GitObjectID" }
        manifest_sha256: { $ref: "#/components/schemas/SHA256" }
        support_level: { $ref: "#/components/schemas/CapabilitySupportLevel" }
        pinned_at: { type: string, format: date-time }
...
    ManualArtifactState:
      type: string
      enum: [pending, ready, failed, stale]
    ManualArtifact:
      type: object
      additionalProperties: false
      required:
        - id
        - canonical_revision_id
        - exact_core_version
        - core_artifact_id
        - config_sha256
        - raw
        - diagnostics
        - state
        - created_at
      properties:
        id: { type: string }
        canonical_revision_id: { type: string }
        exact_core_version: { $ref: "#/components/schemas/ExactCoreVersion" }
        core_artifact_id: { type: string }
        config_sha256: { $ref: "#/components/schemas/SHA256" }
        raw: { type: string }
        diagnostics:
          type: array
          items: {}
        state: { $ref: "#/components/schemas/ManualArtifactState" }
        checked_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
...
    CapabilityDiagnostic:
      type: object
      additionalProperties: false
      required: [severity, code, fact_id, message]
      properties:
        severity: { type: string, const: warning }
        code: { type: string }
        fact_id: { type: string }
        message: { type: string }
    StructuredRenderRequest:
      type: object
      additionalProperties: false
      properties:
        core_version:
          allOf:
            - { $ref: "#/components/schemas/ExactCoreVersion" }
          description: Optional exact version. Omit to use the OS-verified running exact core version; latest is never inferred.
        core_artifact_id: { type: string, minLength: 1, maxLength: 256 }
        allow_compatible:
          type: boolean
          default: false
          description: Explicit acceptance required for compatible_structured capability support.
    StructuredArtifact:
      type: object
      additionalProperties: false
      required:
        - id
        - canonical_revision_id
        - exact_core_version
        - capability_commit
        - capability_digest
        - renderer_version
        - core_artifact_id
        - config_sha256
        - config
        - diagnostics
        - state
      properties:
        id: { type: string }
        canonical_revision_id: { type: string }
        exact_core_version: { $ref: "#/components/schemas/ExactCoreVersion" }
        capability_commit: { $ref: "#/components/schemas/GitObjectID" }
        capability_digest: { $ref: "#/components/schemas/SHA256" }
        renderer_version: { type: string }
        core_artifact_id: { type: string }
        config:
          type: object
          additionalProperties: true
          description: Exact strict sing-box JSON bytes persisted in the startup artifact.
        config_sha256: { $ref: "#/components/schemas/SHA256" }
        diagnostics:
          type: array
          items: { $ref: "#/components/schemas/CapabilityDiagnostic" }
        state: { $ref: "#/components/schemas/ManualArtifactState" }
    StructuredRender:
      type: object
      additionalProperties: false
      required: [resolution, pin, artifact, task]
      properties:
        resolution: { $ref: "#/components/schemas/CoreVersionResolution" }
        pin: { $ref: "#/components/schemas/CapabilityPin" }
        artifact: { $ref: "#/components/schemas/StructuredArtifact" }
        task: { $ref: "#/components/schemas/Task" }
    StartupArtifactSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - kind
        - canonical_revision_id
        - exact_core_version
        - renderer_version
        - core_artifact_id
        - config_sha256
        - diagnostics
        - state
        - created_at
      properties:
        id: { type: string }
        kind: { type: string, enum: [structured, manual] }
        canonical_revision_id: { type: string }
        exact_core_version: { $ref: "#/components/schemas/ExactCoreVersion" }
        capability_commit: { $ref: "#/components/schemas/GitObjectID" }
        capability_digest: { $ref: "#/components/schemas/SHA256" }
        renderer_version: { type: string }
        core_artifact_id: { type: string }
        config_sha256: { $ref: "#/components/schemas/SHA256" }
        diagnostics:
          type: array
          items: {}
        state: { $ref: "#/components/schemas/ManualArtifactState" }
        checked_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    StartupArtifactCursor:
      type: object
      additionalProperties: false
      required: [created_at, id]
      properties:
        created_at: { type: string, format: date-time }
        id: { type: string }
    StartupArtifactPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/StartupArtifactSummary" }
        next: { $ref: "#/components/schemas/StartupArtifactCursor" }
    ManualReattachPinEvidence:
      type: object
      additionalProperties: false
      required: [repository, commit_sha, manifest_sha256, support_level]
      properties:
        repository: { type: string, const: rehuony/sing-box-panel }
        commit_sha: { $ref: "#/components/schemas/GitObjectID" }
        manifest_sha256: { $ref: "#/components/schemas/SHA256" }
        support_level: { $ref: "#/components/schemas/CapabilitySupportLevel" }
    ManualReattachEvidence:
      type: object
      additionalProperties: false
      required:
        - startup_artifact_id
        - config_sha256
        - base_revision_id
        - base_revision_sha256
        - current_head_id
        - current_head_sha256
        - exact_core_version
        - core_artifact_id
        - capability
      properties:
        startup_artifact_id: { type: string }
        config_sha256: { $ref: "#/components/schemas/SHA256" }
        base_revision_id: { type: string }
        base_revision_sha256: { $ref: "#/components/schemas/SHA256" }
        current_head_id: { type: string }
        current_head_sha256: { $ref: "#/components/schemas/SHA256" }
        exact_core_version: { $ref: "#/components/schemas/ExactCoreVersion" }
        core_artifact_id: { type: string }
        capability: { $ref: "#/components/schemas/ManualReattachPinEvidence" }
    ReconcileEncodedValue:
      type: object
      additionalProperties: false
      required: [present]
      properties:
        present: { type: boolean }
        value: {}
    ManualReattachConflict:
      type: object
      additionalProperties: false
      required: [path, base, current, manual]
      properties:
        path: { type: string }
        base: { $ref: "#/components/schemas/ReconcileEncodedValue" }
        current: { $ref: "#/components/schemas/ReconcileEncodedValue" }
        manual: { $ref: "#/components/schemas/ReconcileEncodedValue" }
    ManualReattachPreview:
      type: object
      additionalProperties: false
      required: [evidence, base, current, manual, owned_partial, merged, residual_paths, conflicts]
      properties:
        evidence: { $ref: "#/components/schemas/ManualReattachEvidence" }
        base: { $ref: "#/components/schemas/CanonicalSnapshot" }
        current: { $ref: "#/components/schemas/CanonicalSnapshot" }
        manual: { $ref: "#/components/schemas/CanonicalDocument" }
        owned_partial: { type: object, additionalProperties: true }
        merged: { $ref: "#/components/schemas/CanonicalDocument" }
        residual_paths:
          type: array
          items: { type: string }
        conflicts:
          type: array
          items: { $ref: "#/components/schemas/ManualReattachConflict" }
    ManualReattachApplyRequest:
      type: object
      additionalProperties: false
      required: [evidence, decisions]
      properties:
        evidence: { $ref: "#/components/schemas/ManualReattachEvidence" }
        decisions:
          type: object
          additionalProperties:
            type: string
            enum: [current, manual]
    ManualReattachSave:
      type: object
      additionalProperties: false
      required: [preview, revision, artifact, task]
      properties:
        preview: { $ref: "#/components/schemas/ManualReattachPreview" }
        revision: { $ref: "#/components/schemas/CanonicalSnapshot" }
        artifact: { $ref: "#/components/schemas/ManualArtifact" }
        task: { $ref: "#/components/schemas/Task" }
...
    TaskStatus:
      type: string
      enum: [queued, running, succeeded, failed, canceled, superseded]
    Task:
      type: object
      additionalProperties: false
      required:
        - id
        - lane
        - kind
        - status
        - generation
        - payload
        - cancel_requested
        - attempt
        - created_at
        - updated_at
      properties:
        id: { type: string }
        idempotency_key: { type: string }
        lane: { type: string, enum: [runtime, maintenance] }
        kind: { type: string }
        status: { $ref: "#/components/schemas/TaskStatus" }
        generation: { type: integer, format: int64, minimum: 0 }
        canonical_revision_id: { type: string }
        startup_artifact_id: { type: string }
        activation_bundle_id: { type: string }
        payload: {}
        result: {}
        failure: {}
        cancel_requested: { type: boolean }
        attempt: { type: integer, minimum: 0 }
        lease_expires_at: { type: string, format: date-time }
        not_before: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
security:
  - sessionCookie: []
  - managementBearer: []